chore: #25 - Rename to ELT pattern
