--- lubrikit/collection/__init__.py
@@ -1,3 +1,0 @@
-from lubrikit.collection.connectors import GoogleDriveAPIConnector, HTTPConnector
-
-__all__ = ["HTTPConnector", "GoogleDriveAPIConnector"]
