--- tests/test_sample.py
@@ -1,2 +1,0 @@
-def test_sample() -> None:
-    assert 1 + 1 == 2
